Compile i18n messages during deploy.
Here we introduce a new post-compile hook to run `manage.py compilemessages`
for each one of our apps. This depends on the presence of the GNU gettext
binaries in the production server. To achieve that, we use our
[heroku-buildpack-gettext](https://github.com/labhackercd/heroku-buildpack-gettext).

See [Heroku's documentation on multiple
buildpacks](https://devcenter.heroku.com/articles/using-multiple-buildpacks-for-an-app)
for more information about using multiple buildpacks.

In our specific case, we need to make sure the order of buildpacks is:

1. heroku-buildpack-gettext
2. heroku-buildpack-python

This way gettext will be available during python's compile, which is where our
post-compile hook will run.